Issue#225: EnumConverterProvider, JacksonJsonProvider, LocalDateTimeProvider, StringToContactTypeConverter moved to provider package
