+++ home-application/src/main/java/com/softserveinc/ita/homeproject/application/provider/StringToContactTypeConverter.java
@@ -1,4 +1,4 @@
-package com.softserveinc.ita.homeproject.application.config;
+package com.softserveinc.ita.homeproject.application.provider;
 
 import javax.ws.rs.ext.Provider;
